pyproject.toml
[project]
name = "wecompany"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.0",
    "bleach>=6.3.0",
    "django>=6.0.4",
    "django-q2>=1.9.0",
    "django-ratelimit>=4.1.0",
    "markdown>=3.10.2",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "pygithub>=2.9.1",
    "python-dotenv>=1.2.2",
]
